feat(slack): improve slack weekly digest, add public holidays info
* phpstan updates

+++ app/src/Infrastructure/Doctrine/Repository/InvitationRepository.php
@@ -31,6 +31,10 @@ class InvitationRepository extends ServiceEntityRepository implements Invitation
     {
         $invitation = $this->findOneBy(['token' => $invitationDTO->token]);
 
+        if (!$invitation instanceof Invitation) {
+            return;
+        }
+
         $em = $this->getEntityManager();
 
         $em->remove($invitation);
